model-view separation
"Although not strictly necessary, separating the view from the model is a good practice that was originally advocated by Smalltalk developers. Highlighting the demarcation between object mode land user interface encourages functionality to be added to the most appropriate class. The Model schema contains classes that encapsulate the database behavior while the View schema has classes that define the user interface. This allows multiple interfaces to be built that use the same database definition. It also allows a development team to have members working in their various areas of expertise"